+++ springql-core/src/api/spring_source_row.rs
@@ -1,5 +1,9 @@
 // This file is part of https://github.com/SpringQL/SpringQL which is licensed under MIT OR Apache-2.0. See file LICENSE-MIT or LICENSE-APACHE for full license details.
 
+mod spring_source_row_builder;
+
+pub use spring_source_row_builder::SpringSourceRowBuilder;
+
 use crate::{
     api::error::Result,
     stream_engine::autonomous_executor::{SchemalessRow, SourceRow},